LayoutTests/imported/w3c/web-platform-tests/content-security-policy/navigate-to/form-action/form-action-blocks-navigate-to-allows.sub-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'navigate-to'.

CONSOLE MESSAGE: Refused to load http://localhost:8800/content-security-policy/navigate-to/support/post_message_to_frame_owner.html?dummy= because it does not appear in the form-action directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

NOTRUN Test that form-action overrides navigate-to when present.
